chore: updates per review - Renaming ui_utils.py to core_ui.py - Minor style fixes and unused imports removal
