+++ tests/test_core.py
@@ -3,7 +3,7 @@ import math
 from typing import Dict
 
 import numpy as np
-from streamsync.core import (BytesWrapper, ComponentTree, Evaluator, EventDeserialiser,
+from streamsync.core import (BytesWrapper, Evaluator, EventDeserialiser,
                              FileWrapper, SessionManager, StateProxy, StateSerialiser, StateSerialiserException, StreamsyncState)
 import streamsync as ss
 from streamsync.ss_types import StreamsyncEvent
